modularized lhc module
